project_conecta_saude/service_llm/requirements.txt
fastapi
uvicorn[standard]
pydantic

# --- Pacotes LangChain ---
langchain
langchain-google-genai  # Integração do Gemini

# --- Pacotes RAG (PDFs) ---
pypdf                   # Para ler os arquivos PDF
faiss-cpu               # Para o banco de dados vetorial (Vector Store)

# Para carregar a API Key do .env
python-dotenv

langchain-community
langchain-text-splitters

sentence-transformers
langchain-huggingface